// src/models/product-model.ts
export interface ServerResponse {
  products: Product[],
  total: number,
  skip: number,
  limit: number,
}

export interface Product {
  id: number;
  title: string;
  description: string;

  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;

  brand: string;
  category: string;

  thumbnail: string;
  images: string[];

  tags?: string[];
  warrantyInformation?: string;
  shippingInformation?: string;
}